add the realestate-admin module
