+++ src/app/app.module.ts
@@ -1,15 +1,24 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { NgZorroAntdModule } from 'ng-zorro-antd';
-
+import { HttpClientModule } from '@angular/common/http';
 import { RoutesModule } from './routes/routes.module';
 import { AppComponent } from './app.component';
+import { NgZorroAntdModule, NZ_I18N, zh_CN } from 'ng-zorro-antd';
+/** 配置 angular i18n **/
+import { registerLocaleData } from '@angular/common';
+import zh from '@angular/common/locales/zh';
+registerLocaleData(zh);
 
 @NgModule({
+  imports: [
+    BrowserModule,
+    HttpClientModule,
+    NgZorroAntdModule,
+    RoutesModule
+  ],
   declarations: [AppComponent],
-  imports: [BrowserModule, NgZorroAntdModule.forRoot(), RoutesModule],
-  bootstrap: [AppComponent],
   providers: [
+    { provide: NZ_I18N, useValue: zh_CN }
     // StartupService,
     // {
     //   provide: APP_INITIALIZER,
@@ -17,6 +26,7 @@ import { AppComponent } from './app.component';
     //   deps: [StartupService],
     //   multi: true
     // }
-  ]
+  ],
+  bootstrap: [AppComponent],
 })
-export class AppModule {}
+export class AppModule { }
