--- src/app/business/dashboard/dashboard.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { DashboardRoutingModule } from './dashboard.routing.module';
-
-import { DashboardComponent } from './dashboard.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    DashboardRoutingModule
-  ],
-  declarations: [DashboardComponent]
-})
-export class DashboardModule { }
