+++ src/app/business/realestate-admin/realestate-admin.component.html
@@ -1,0 +1,3 @@
+<p>
+  realestate-admin works!
+</p>
