+++ src/app/business/realestate-admin/realestate-admin.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-realestate-admin',
+  templateUrl: './realestate-admin.component.html',
+  styleUrls: ['./realestate-admin.component.scss']
+})
+export class RealestateAdminComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
