+++ src/app/business/realestate-admin/realestate-admin.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RealestateAdminComponent } from './realestate-admin.component';
+import { RealestateAdminRoutesModule } from './realestate-admin.routing';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    RealestateAdminRoutesModule
+  ],
+  declarations: [RealestateAdminComponent]
+})
+export class RealestateAdminModule { }
